more user messages

+++ test-it/src/app/tests/components/test-detail/test-detail.component.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Observable, map, mergeMap, switchMap } from 'rxjs';
+import { DialogComponent } from 'src/app/dialog/dialog.component';
 import { AuthService } from 'src/app/services/auth.service';
 import { ITest, TestsService } from 'src/app/services/tests.service';
 
@@ -12,12 +14,14 @@ import { ITest, TestsService } from 'src/app/services/tests.service';
 export class TestDetailComponent {
   test$: Observable<ITest> | undefined = undefined;
   isTestOwners = false;
+  title: string | undefined = undefined;
 
   constructor(
     private testsService: TestsService,
     private activatedRoute: ActivatedRoute,
     private authService: AuthService,
-    private router: Router
+    private router: Router,
+    public dialog: MatDialog
   ) {}
 
   ngOnInit() {
@@ -27,11 +31,23 @@ export class TestDetailComponent {
 
     this.test$?.subscribe((test) => {
       this.isTestOwners = test.ownerId == this.authService.user?._id;
+      this.title = test.title;
     });
   }
 
   deleteTest(test: ITest) {
-    this.testsService.deleteTest(test?._id || '');
-    this.router.navigate(['/tests']);
+    const dialogRef = this.dialog.open(DialogComponent, {
+      data: {
+        title: `Are you sure you want to delete ${this.title}`,
+        confirmText: 'DELETE',
+      },
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.testsService.deleteTest(test?._id || '');
+        this.router.navigate(['/tests']);
+      }
+    });
   }
 }
